example version 3.4(8)/ we can draw and show the lines in line_drawing.dart/ we can see the rotation result both raw and processed/ better slope calculation/ better lining algorithm

+++ lib/classes.dart
@@ -64,8 +64,8 @@ class CornerPoint {
   CornerPoint({this.x, this.y});
 
   CornerPoint.fromJson(Map<String, dynamic> json) {
-    x = json['x'];
-    y = json['y'];
+    x = json['x'].toDouble();
+    y = json['y'].toDouble();
   }
 
   Map<String, dynamic> toJson() {
